Split settings sheet parity gates
